requirements.txt
altgraph==0.17.2
black==22.1.0
click==8.0.4
colorama==0.4.4
defusedxml==0.7.1
Flask==2.0.3
fpdf2==2.5.1
future==0.18.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
pathspec==0.9.0
pefile==2021.9.3
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.19.4
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pywin32-ctypes==0.2.0
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
